Abstract PEMDAS logic to operate()
Before, you'd have to call the specific functions to do any PEMDAS sort
of calculations, and now you only have to call operate, and it'll pick
up based on the operator what you want to do.

Reordered the function calls in to the operate() being at top of file,
and having the prior alphabetic ordering of functions to PEMDAS order.

Any other functions that may be necessary in the future will be in
alphabetical order below the operate and PEMDAS functions.