Set up Jest for Test Driven Development
Why: Annoys me to just sanity check all of my JS logic once I think it
is done by using the browser console, that's manual and a waste of life.
I'd rather automate the testing of sanity checks, so I can avoid more
pain from bad code later on.

Why the if statement for the module.exports?
The in-browser JS console was throwing a fit, that if statement makes it
stop complaining about it. I posted the Reddit URL at the relevant point
in the code, but incase that gets removed. I relabeled it with a bit.ly
to make what was the point of the URL more clear for me.

URL: https://bit.ly/fix_for_referenceError_module_is_not_defined