Remove & revamp pseudo-code comments
Before, they were helpful while creating & debugging the functions, but
keeping them in when the functions have been sufficiently tested & seem
to work as intended is just a maintenance risk, as there's no given that
the comments will stay up-to-date if future fixes need to be made.

Revamped pseudo-code comments on setOperand() to leave an inline doc
of how the outcomes of certain conditionals are the same, but still
thinking about if I want to rework the code that way or not.