Remove remaining pseudo-code comments
Decided against revamping the setOperand() code to the pseudo-code,
because I felt like it was harder to figure out the big-picture logic
of the function when it was done that way.

Less lines of code is meant to be a method to improve readability, but
refactoring to deliberately get less lines of code at the trade-off of
worsening readability is not worth it to me.