Revamp calculator UI for display & buttons
Before, there was no clear start or end to the display or the calculator
itself. Now, there's a coloring of the display to indicate where inputs
will go, and there's a horizontal line at the bottom to make it more
clear where the calculator UI ends.

The horizontal is strictly for aesthetics, so I did not use the <hr> tag
as that is intended for semantic HTML in HTML5.